Create Protohack-System.js
I named it like this because it sounds cool. Also this took too much time for me to admit. Im not sure if it works anymore due to the updates. This essentially makes one server's job to constantly hack/grow/weaken a server but schedules the worker scripts to  start after a specified time so all of them finish the hack/grow/weaken all at once. Plus calculates the RAM needed for each thread of hack/grow/weaken and checks if it can be applied on a server.
this was such a mess