Create LazyShotgunhack.js
This was me giving up on working on the same type of coding. Also I called it a shotgun because it scans all servers and then if it can it will apply protohack deploy for each one. All these worker scripts are all housed inside of a big server with a lot of ram. Also this was me experimenting with importing and library scripts.